Pager moved to common
